Allow YAML metadata header in markdown files
Allow parsing YAML metadata from a header section in markdown files

Update `Note` struct to hold raw markdown and tags and title parsed from the header, or falling back to a title built from the filepath.

Add `Backlink` to construct backlinks which store the text of the link, the expected path and slug based on that title, and whether or not the expected file exists.

On the note page view, display all tags as links in the left nav, concatenating tags from the header with tags in the markdown body

- Add `Plug.Logger` to `XettelkastenServer.Router`
- Add helper methods to `XettelkastenServer.TextHelpers` for common conversions between slugs, text, and paths