Provide bidirectional backlink catalogs when viewing a note
Adds the ability to generate a list of backlinks inside a note, and
to generate a list of notes that link to a specific note.

Both these lists are displayed in the side (top on narrow screens) nav
while viewing a note.

This feature removes the ability to link to a note via its title
directly. Doing so created a circular logic where to define a backlink
the attached note needed to be loaded, but to load that note it had
to search for backlinks, etc.

NB. There may be a way to allow backlinks to display titles, but they would
still have to be detected/defined by the linked path.