Add GitWatcher to check for and commit changed files under the notes directory
This functionality runs `git status` against the configured notes
directory and adds, commits, and pushes any changes it finds.
It can be run ad hoc in an IEX session, or opted into via app configuration
to run every 10 minutes while the server is running.

The README has been updated to explain what it expects and how to activate it.

* To improve the local feedback loop, don't run tests with a delay set locally, but make sure they run in CI